Modularize: Export calendar() without HebrewCalendar
This only has modest savings, because most of the code comes from this very function.

This is not worth splitting; instead, we should redo the API first.
