+++ size-demo/sizes.md
@@ -2,8 +2,8 @@
 
 - `getHoliday.mjs`: 129 kB
 - `getHoliday.min.mjs`: 41.9 kB
-- `parshiyot.mjs`: 407 kB
-- `parshiyot.min.mjs`: 144 kB
+- `parshiyot.mjs`: 385 kB
+- `parshiyot.min.mjs`: 140 kB
 - `sedra.mjs`: 85.7 kB
 - `sedra.min.mjs`: 29.9 kB
 - `tachanun.mjs`: 55.5 kB
