+++ src/hebcal.ts
@@ -19,26 +19,10 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  getBirthdayHD,
-  getYahrzeitHD,
-  abs2greg,
-  months,
-  HDate,
-  Locale,
-} from '@hebcal/hdate';
+import {getBirthdayHD, getYahrzeitHD, HDate, months} from '@hebcal/hdate';
 import './locale'; // Adds Hebrew and Ashkenazic translations
-import {CalOptions, DailyLearningValue} from './CalOptions';
+import {CalOptions} from './CalOptions';
 import {version as pkgVersion} from './pkgVersion';
-import {DailyLearning} from './DailyLearning';
-import {HebrewDateEvent} from './HebrewDateEvent';
-import {ParshaEvent} from './ParshaEvent';
-import {
-  makeCandleEvent,
-  makeFastStartEnd,
-  makeWeekdayChanukahCandleLighting,
-} from './candles';
-import {TimedEvent, HavdalahEvent} from './TimedEvent';
 import {Event, flags} from './event';
 import {Sedra, getSedra} from './sedra';
 import {hallel_} from './hallel';
@@ -48,313 +32,10 @@ import {
   holidayFilter,
   HolidayYearMap,
 } from './holidays';
-import {MevarchimChodeshEvent} from './MevarchimChodeshEvent';
 import {HolidayEvent} from './HolidayEvent';
-import {Location} from './location';
-import {Molad, MoladEvent} from './molad';
-import {OmerEvent} from './omer';
 import {reformatTimeStr} from './reformatTimeStr';
 import {TachanunResult, tachanun} from './tachanun';
-import {Zmanim} from './zmanim';
-import {getStartAndEnd} from './getStartAndEnd';
-
-const FRI = 5;
-const SAT = 6;
-
-const NISAN = months.NISAN;
-const SIVAN = months.SIVAN;
-const ELUL = months.ELUL;
-const TISHREI = months.TISHREI;
-
-const LIGHT_CANDLES = flags.LIGHT_CANDLES;
-const YOM_TOV_ENDS = flags.YOM_TOV_ENDS;
-const CHUL_ONLY = flags.CHUL_ONLY;
-const IL_ONLY = flags.IL_ONLY;
-const LIGHT_CANDLES_TZEIS = flags.LIGHT_CANDLES_TZEIS;
-const CHANUKAH_CANDLES = flags.CHANUKAH_CANDLES;
-const MINOR_FAST = flags.MINOR_FAST;
-const SPECIAL_SHABBAT = flags.SPECIAL_SHABBAT;
-const MODERN_HOLIDAY = flags.MODERN_HOLIDAY;
-const MAJOR_FAST = flags.MAJOR_FAST;
-const ROSH_CHODESH = flags.ROSH_CHODESH;
-const PARSHA_HASHAVUA = flags.PARSHA_HASHAVUA;
-const DAF_YOMI = flags.DAF_YOMI;
-const MISHNA_YOMI = flags.MISHNA_YOMI;
-const NACH_YOMI = flags.NACH_YOMI;
-const YERUSHALMI_YOMI = flags.YERUSHALMI_YOMI;
-const OMER_COUNT = flags.OMER_COUNT;
-const SHABBAT_MEVARCHIM = flags.SHABBAT_MEVARCHIM;
-const MINOR_HOLIDAY = flags.MINOR_HOLIDAY;
-const EREV = flags.EREV;
-const CHOL_HAMOED = flags.CHOL_HAMOED;
-const YOM_KIPPUR_KATAN = flags.YOM_KIPPUR_KATAN;
-
-type StringIntMap = {
-  [x: string]: number;
-};
-
-const unrecognizedAlreadyWarned = new Set<string>();
-const RECOGNIZED_OPTIONS: StringIntMap = {
-  location: 1,
-  year: 1,
-  isHebrewYear: 1,
-  month: 1,
-  numYears: 1,
-  start: 1,
-  end: 1,
-  candlelighting: 1,
-  candleLightingMins: 1,
-  havdalahMins: 1,
-  havdalahDeg: 1,
-  fastEndDeg: 1,
-  sedrot: 1,
-  il: 1,
-  noMinorFast: 1,
-  noModern: 1,
-  shabbatMevarchim: 1,
-  noRoshChodesh: 1,
-  noSpecialShabbat: 1,
-  noHolidays: 1,
-  omer: 1,
-  molad: 1,
-  ashkenazi: 1,
-  locale: 1,
-  addHebrewDates: 1,
-  addHebrewDatesForEvents: 1,
-  mask: 1,
-  yomKippurKatan: 1,
-  hour12: 1,
-  dailyLearning: 1,
-  useElevation: 1,
-} as const satisfies Record<keyof CalOptions, 1>;
-
-/**
- * @private
- */
-function warnUnrecognizedOptions(options: CalOptions) {
-  for (const k of Object.keys(options)) {
-    if (
-      typeof RECOGNIZED_OPTIONS[k] === 'undefined' &&
-      !unrecognizedAlreadyWarned.has(k)
-    ) {
-      console.warn(`Ignoring unrecognized HebrewCalendar option: ${k}`);
-      unrecognizedAlreadyWarned.add(k);
-    }
-  }
-}
-
-const israelCityOffset: StringIntMap = {
-  Jerusalem: 40,
-  Haifa: 30,
-  "Zikhron Ya'aqov": 30,
-  "Zikhron Ya'akov": 30,
-  'Zikhron Yaakov': 30,
-  "Zichron Ya'akov": 30,
-  'Zichron Yaakov': 30,
-} as const;
-
-const geoIdCandleOffset: StringIntMap = {
-  '281184': 40, // Jerusalem
-  '294801': 30, // Haifa
-  '293067': 30, // Zikhron Yaakov
-} as const;
-
-/**
- * @private
- * @constant
- * This calculation is based on the position of the sun 36 minutes after sunset in Jerusalem
- * around the equinox / equilux, which is 8.5° below geometric zenith.
- * The Ohr Meir considers this the time that 3 small stars are visible,
- * which is later than the required 3 medium stars.
- * @see {https://kosherjava.com/zmanim/docs/api/com/kosherjava/zmanim/ZmanimCalendar.html#ZENITH_8_POINT_5}
- */
-const TZEIT_3SMALL_STARS = 8.5;
-
-/**
- * @private
- * @constant
- * This calculation is based on observation of 3 medium sized stars by Dr. Baruch Cohen
- * in his calendar published in in 1899 in Strasbourg, France.
- * This calculates to 7.0833333° below geometric zenith.
- * @see {https://kosherjava.com/zmanim/docs/api/com/kosherjava/zmanim/ComplexZmanimCalendar.html#ZENITH_7_POINT_083}
- */
-const TZEIT_3MEDIUM_STARS = 7.0833333;
-
-/**
- * Modifies options in-place
- * @private
- */
-function checkCandleOptions(options: CalOptions) {
-  if (!options.candlelighting) {
-    return;
-  }
-  const location = options.location;
-  if (typeof location === 'undefined' || !(location instanceof Location)) {
-    throw new TypeError(
-      'options.candlelighting requires valid options.location'
-    );
-  }
-  if (
-    typeof options.havdalahMins === 'number' &&
-    typeof options.havdalahDeg === 'number'
-  ) {
-    throw new TypeError(
-      'options.havdalahMins and options.havdalahDeg are mutually exclusive'
-    );
-  }
-
-  let min = Number(options.candleLightingMins) || 18;
-  if (location.getIsrael() && Math.abs(min) === 18) {
-    min = overrideIsraelCandleMins(location, min);
-  }
-  options.candleLightingMins = -1 * Math.abs(min);
-
-  if (typeof options.havdalahMins === 'number') {
-    options.havdalahMins = Math.abs(options.havdalahMins);
-  } else if (typeof options.havdalahDeg === 'number') {
-    options.havdalahDeg = Math.abs(options.havdalahDeg);
-  } else {
-    options.havdalahDeg = TZEIT_3SMALL_STARS;
-  }
-  if (typeof options.fastEndDeg !== 'number') {
-    options.fastEndDeg = TZEIT_3MEDIUM_STARS;
-  }
-}
-
-function overrideIsraelCandleMins(location: Location, min: number) {
-  const geoid = location.getGeoId();
-  if (geoid) {
-    const offset = geoIdCandleOffset[geoid];
-    if (typeof offset === 'number') {
-      return offset;
-    }
-  }
-  const shortName = location.getShortName();
-  if (shortName) {
-    const offset = israelCityOffset[shortName];
-    if (typeof offset === 'number') {
-      return offset;
-    }
-  }
-  return min;
-}
-
-/**
- * Mask to filter Holiday array
- * @private
- */
-function getMaskFromOptions(options: CalOptions): number {
-  if (typeof options.mask === 'number') {
-    return setOptionsFromMask(options);
-  }
-  const il = options.il || options.location?.getIsrael() || false;
-  let mask = 0;
-
-  // default options
-  if (!options.noHolidays) {
-    mask |=
-      ROSH_CHODESH |
-      YOM_TOV_ENDS |
-      MINOR_FAST |
-      SPECIAL_SHABBAT |
-      MODERN_HOLIDAY |
-      MAJOR_FAST |
-      MINOR_HOLIDAY |
-      EREV |
-      CHOL_HAMOED |
-      LIGHT_CANDLES |
-      LIGHT_CANDLES_TZEIS |
-      CHANUKAH_CANDLES;
-  }
-  if (options.candlelighting) {
-    mask |= LIGHT_CANDLES | LIGHT_CANDLES_TZEIS | YOM_TOV_ENDS;
-  }
-  // suppression of defaults
-  if (options.noRoshChodesh) {
-    mask &= ~ROSH_CHODESH;
-  }
-  if (options.noModern) {
-    mask &= ~MODERN_HOLIDAY;
-  }
-  if (options.noMinorFast) {
-    mask &= ~MINOR_FAST;
-  }
-  if (options.noSpecialShabbat) {
-    mask &= ~SPECIAL_SHABBAT;
-    mask &= ~SHABBAT_MEVARCHIM;
-  }
-  if (il) {
-    mask |= IL_ONLY;
-  } else {
-    mask |= CHUL_ONLY;
-  }
-  // non-default options
-  if (options.sedrot) {
-    mask |= PARSHA_HASHAVUA;
-  }
-  if (options.omer) {
-    mask |= OMER_COUNT;
-  }
-  if (options.shabbatMevarchim) {
-    mask |= SHABBAT_MEVARCHIM;
-  }
-  if (options.yomKippurKatan) {
-    mask |= YOM_KIPPUR_KATAN;
-  }
-  if (options.dailyLearning) {
-    const dailyLearning = options.dailyLearning;
-    if (dailyLearning.dafYomi) {
-      mask |= DAF_YOMI;
-    }
-    if (dailyLearning.mishnaYomi) {
-      mask |= MISHNA_YOMI;
-    }
-    if (dailyLearning.nachYomi) {
-      mask |= NACH_YOMI;
-    }
-    if (dailyLearning.yerushalmi) {
-      mask |= YERUSHALMI_YOMI;
-    }
-  }
-
-  return mask;
-}
-
-const MASK_LIGHT_CANDLES =
-  LIGHT_CANDLES | LIGHT_CANDLES_TZEIS | CHANUKAH_CANDLES | YOM_TOV_ENDS;
-
-const defaultLocation = new Location(0, 0, false, 'UTC');
-
-/**
- * @private
- */
-function setOptionsFromMask(options: CalOptions): number {
-  const m = options.mask || 0;
-  if (m & ROSH_CHODESH) delete options.noRoshChodesh;
-  if (m & MODERN_HOLIDAY) delete options.noModern;
-  if (m & MINOR_FAST) delete options.noMinorFast;
-  if (m & SPECIAL_SHABBAT) delete options.noSpecialShabbat;
-  if (m & PARSHA_HASHAVUA) options.sedrot = true;
-  if (m & (DAF_YOMI | MISHNA_YOMI | NACH_YOMI | YERUSHALMI_YOMI)) {
-    options.dailyLearning = options.dailyLearning || {};
-    if (m & DAF_YOMI) {
-      options.dailyLearning.dafYomi = true;
-    }
-    if (m & MISHNA_YOMI) {
-      options.dailyLearning.mishnaYomi = true;
-    }
-    if (m & NACH_YOMI) {
-      options.dailyLearning.nachYomi = true;
-    }
-    if (m & YERUSHALMI_YOMI) {
-      options.dailyLearning.yerushalmi = 1;
-    }
-  }
-  if (m & OMER_COUNT) options.omer = true;
-  if (m & SHABBAT_MEVARCHIM) options.shabbatMevarchim = true;
-  if (m & YOM_KIPPUR_KATAN) options.yomKippurKatan = true;
-  return m;
-}
+import {calendar} from './calendar';
 
 /**
  * HebrewCalendar is the main interface to the `@hebcal/core` library.
@@ -472,131 +153,7 @@ export class HebrewCalendar {
    * }
    */
   static calendar(options: CalOptions = {}): Event[] {
-    options = {...options}; // so we can modify freely
-    checkCandleOptions(options);
-    const location = (options.location = options.location || defaultLocation);
-    const il = (options.il = options.il || location.getIsrael() || false);
-    const hasUserMask = typeof options.mask === 'number';
-    options.mask = getMaskFromOptions(options);
-    if (options.ashkenazi || options.locale) {
-      if (options.locale && typeof options.locale !== 'string') {
-        throw new TypeError(`Invalid options.locale: ${options.locale}`);
-      }
-      const locale = options.ashkenazi
-        ? 'ashkenazi'
-        : (options.locale as string);
-      const translationObj = Locale.useLocale(locale);
-      if (!translationObj) {
-        throw new TypeError(
-          `Locale '${locale}' not found; did you forget to import @hebcal/locales?`
-        );
-      }
-    } else {
-      Locale.useLocale('en');
-    }
-
-    const evts: Event[] = [];
-    let sedra: Sedra | undefined;
-    let holidaysYear: HolidayYearMap | undefined;
-    let beginOmer = -1;
-    let endOmer = -1;
-    let currentYear = -1;
-    const startAndEnd = getStartAndEnd(options);
-    warnUnrecognizedOptions(options);
-    const startAbs = startAndEnd[0];
-    const endAbs = startAndEnd[1];
-    const startGreg = abs2greg(startAbs);
-    if (startGreg.getFullYear() < 100) {
-      options.candlelighting = false;
-    }
-    for (let abs = startAbs; abs <= endAbs; abs++) {
-      const hd = new HDate(abs);
-      const hyear = hd.getFullYear();
-      if (hyear !== currentYear) {
-        currentYear = hyear;
-        holidaysYear = getHolidaysForYear_(currentYear);
-        if (options.sedrot) {
-          sedra = getSedra(currentYear, il);
-        }
-        if (options.omer) {
-          beginOmer = HDate.hebrew2abs(currentYear, NISAN, 16);
-          endOmer = HDate.hebrew2abs(currentYear, SIVAN, 5);
-        }
-      }
-      const prevEventsLength = evts.length;
-      const dow = hd.getDay();
-      const isFriday = dow === FRI;
-      const isSaturday = dow === SAT;
-      let candlesEv: TimedEvent | undefined;
-      const holidays = holidaysYear!.get(hd.toString()) || [];
-      for (const ev of holidays) {
-        candlesEv = appendHolidayAndRelated(
-          candlesEv,
-          evts,
-          ev,
-          options,
-          isFriday,
-          isSaturday,
-          hasUserMask
-        );
-      }
-      if (options.sedrot && isSaturday) {
-        const parsha0 = sedra!.lookup(abs);
-        if (!parsha0.chag) {
-          evts.push(new ParshaEvent(hd, parsha0.parsha, il, parsha0.num));
-        }
-      }
-      const dailyLearning = options.dailyLearning;
-      if (typeof dailyLearning === 'object') {
-        const events = makeDailyLearning(hd, dailyLearning, il);
-        evts.push(...events);
-      }
-      if (options.omer && abs >= beginOmer && abs <= endOmer) {
-        const omer = abs - beginOmer + 1;
-        const omerEv = makeOmerEvent(hd, omer, options);
-        evts.push(omerEv);
-      }
-      if (isSaturday && (options.molad || options.shabbatMevarchim)) {
-        const events = makeMoladAndMevarchimChodesh(hd, options);
-        evts.push(...events);
-      }
-      if (!candlesEv && options.candlelighting && (isFriday || isSaturday)) {
-        candlesEv = makeCandleEvent(
-          undefined,
-          hd,
-          options,
-          isFriday,
-          isSaturday
-        );
-        if (isFriday && candlesEv && sedra) {
-          const parsha = sedra.lookup(abs);
-          const pe = new ParshaEvent(hd.next(), parsha.parsha, il, parsha.num);
-          candlesEv.memo = pe.render(options.locale);
-        }
-      }
-      // suppress Havdalah when options.havdalahMins=0 or options.havdalahDeg=0
-      if (
-        candlesEv instanceof HavdalahEvent &&
-        (options.havdalahMins === 0 || options.havdalahDeg === 0)
-      ) {
-        candlesEv = undefined;
-      }
-      if (candlesEv) {
-        evts.push(candlesEv);
-      }
-      if (
-        options.addHebrewDates ||
-        (options.addHebrewDatesForEvents && prevEventsLength !== evts.length)
-      ) {
-        const e2 = new HebrewDateEvent(hd);
-        if (prevEventsLength === evts.length) {
-          evts.push(e2);
-        } else {
-          evts.splice(prevEventsLength, 0, e2);
-        }
-      }
-    }
-    return evts;
+    return calendar(options);
   }
 
   /**
@@ -695,8 +252,8 @@ export class HebrewCalendar {
    */
   static getHolidaysForYearArray(year: number, il: boolean): HolidayEvent[] {
     const yearMap = getHolidaysForYear_(year);
-    const startAbs = HDate.hebrew2abs(year, TISHREI, 1);
-    const endAbs = HDate.hebrew2abs(year + 1, TISHREI, 1) - 1;
+    const startAbs = HDate.hebrew2abs(year, months.TISHREI, 1);
+    const endAbs = HDate.hebrew2abs(year + 1, months.TISHREI, 1) - 1;
     let events: HolidayEvent[] = [];
     for (let absDt = startAbs; absDt <= endAbs; absDt++) {
       const hd = new HDate(absDt);
@@ -817,129 +374,4 @@ function isChag(date: HDate, il: boolean): boolean {
   const events = getHolidaysOnDate(date, il) || [];
   const chag = events.filter(ev => ev.getFlags() & flags.CHAG);
   return chag.length !== 0;
-}
-
-/**
- * Appends the Event `ev` to the `events` array. Also may add related
- * timed events like candle-lighting or fast start/end
- * @private
- */
-function appendHolidayAndRelated(
-  candlesEv: TimedEvent | undefined,
-  events: Event[],
-  ev: HolidayEvent,
-  options: CalOptions,
-  isFriday: boolean,
-  isSaturday: boolean,
-  hasUserMask: boolean
-): TimedEvent | undefined {
-  const il = options.il || false;
-  if (!ev.observedIn(il)) {
-    return candlesEv; // holiday isn't observed here; bail out early
-  }
-  const eFlags = ev.getFlags();
-  if (
-    (!options.yomKippurKatan && eFlags & YOM_KIPPUR_KATAN) ||
-    (options.noModern && eFlags & MODERN_HOLIDAY)
-  ) {
-    return candlesEv; // bail out early
-  }
-  const isMajorFast = Boolean(eFlags & MAJOR_FAST);
-  const isMinorFast = Boolean(eFlags & MINOR_FAST);
-  if (options.candlelighting && (isMajorFast || isMinorFast)) {
-    ev = makeFastStartEnd(ev, options);
-    if (
-      ev.startEvent &&
-      (isMajorFast || (isMinorFast && !options.noMinorFast))
-    ) {
-      events.push(ev.startEvent);
-    }
-  }
-  if (eFlags & Number(options.mask) || (!eFlags && !hasUserMask)) {
-    if (options.candlelighting && eFlags & MASK_LIGHT_CANDLES) {
-      const hd = ev.getDate();
-      candlesEv = makeCandleEvent(ev, hd, options, isFriday, isSaturday);
-      if (eFlags & CHANUKAH_CANDLES && candlesEv && !options.noHolidays) {
-        // Replace Chanukah event with a clone that includes candle lighting time.
-        // For clarity, allow a "duplicate" candle lighting event to remain for Shabbat
-        const chanukahEv = makeWeekdayChanukahCandleLighting(ev, hd, options);
-        if (chanukahEv) {
-          if (isFriday || isSaturday) {
-            chanukahEv.eventTime = candlesEv.eventTime;
-            chanukahEv.eventTimeStr = candlesEv.eventTimeStr;
-          }
-          ev = chanukahEv;
-        }
-        candlesEv = undefined;
-      }
-    }
-    if (
-      !options.noHolidays ||
-      (options.yomKippurKatan && eFlags & YOM_KIPPUR_KATAN)
-    ) {
-      events.push(ev); // the original event itself
-    }
-  }
-  if (ev.endEvent && (isMajorFast || (isMinorFast && !options.noMinorFast))) {
-    events.push(ev.endEvent);
-  }
-  return candlesEv;
-}
-
-function makeMoladAndMevarchimChodesh(hd: HDate, options: CalOptions): Event[] {
-  const evts: Event[] = [];
-  const hmonth = hd.getMonth();
-  const hdate = hd.getDate();
-  if (hmonth !== ELUL && hdate >= 23 && hdate <= 29) {
-    const hyear = hd.getFullYear();
-    const monNext = hmonth === HDate.monthsInYear(hyear) ? NISAN : hmonth + 1;
-    if (options.molad) {
-      evts.push(new MoladEvent(hd, hyear, monNext, options));
-    }
-    if (options.shabbatMevarchim) {
-      const nextMonthName = HDate.getMonthName(monNext, hyear);
-      const molad = new Molad(hyear, monNext);
-      const memo = molad.render(options.locale || 'en', options);
-      evts.push(new MevarchimChodeshEvent(hd, nextMonthName, memo));
-    }
-  }
-  return evts;
-}
-
-function dailyLearningName(key: string, val: DailyLearningValue): string {
-  if (key === 'yerushalmi') {
-    return val === 2 ? 'yerushalmi-schottenstein' : 'yerushalmi-vilna';
-  }
-  return key;
-}
-
-function makeDailyLearning(
-  hd: HDate,
-  dailyLearning: {[x: string]: DailyLearningValue},
-  il: boolean
-): Event[] {
-  const evts: Event[] = [];
-  for (const [key, val] of Object.entries(dailyLearning)) {
-    if (val) {
-      const name = dailyLearningName(key, val);
-      const learningEv = DailyLearning.lookup(name, hd, il);
-      if (learningEv) {
-        evts.push(learningEv);
-      }
-    }
-  }
-  return evts;
-}
-
-function makeOmerEvent(hd: HDate, omerDay: number, options: CalOptions) {
-  const omerEv = new OmerEvent(hd, omerDay);
-  if (options.candlelighting) {
-    const location = options.location!;
-    const zmanim = new Zmanim(location, hd.prev(), false);
-    const tzeit = zmanim.tzeit(7.0833);
-    if (!isNaN(tzeit.getTime())) {
-      omerEv.alarm = tzeit;
-    }
-  }
-  return omerEv;
 }
